requirements.txt
streamlit
pandas
numpy
pathlib
keras
Pillow
tensorflow
